requirements/ci_requirements.txt
pyyaml >= 5.3
tqdm
pylint
pytest
click == 8.1.3
mindspore == 2.0.0
Pillow == 9.5.0
mindformers
requests